requirements.txt
fastapi
uvicorn
requests
docx
pymupdf
sentence-transformers
faiss-cpu
llama-cpp-python
